fix : EventListenerGroup 제거, EventListener 응집화
