+++ src/main/java/persistence/core/MetaModel.java
@@ -5,8 +5,6 @@ import persistence.entity.loader.EntityLoaders;
 import persistence.entity.persister.EntityPersisters;
 import persistence.entity.proxy.EntityProxyFactory;
 import persistence.event.EventListener;
-import persistence.event.EventListenerGroup;
-import persistence.event.EventType;
 
 import java.util.Set;
 
@@ -21,7 +19,7 @@ public interface MetaModel {
 
     EntityProxyFactory getEntityProxyFactory();
 
-    <T extends EventListener> EventListenerGroup<T> getEventListenerGroup(EventType<T> type);
+    EventListener getEventListener();
 
     ActionQueue getActionQueue();
 }
