--- src/main/java/persistence/event/DeleteEventListener.java
@@ -1,5 +1,0 @@
-package persistence.event;
-
-public interface DeleteEventListener extends EventListener {
-    void onDelete(final DeleteEvent deleteEvent);
-}
