--- src/main/java/persistence/event/LoadEventListener.java
@@ -1,5 +1,0 @@
-package persistence.event;
-
-public interface LoadEventListener extends EventListener {
-    <T> T onLoad(LoadEvent<T> loadEvent);
-}
