--- src/main/java/persistence/event/MergeEventListener.java
@@ -1,5 +1,0 @@
-package persistence.event;
-
-public interface MergeEventListener extends EventListener {
-    void onMerge(final MergeEvent mergeEvent);
-}
