--- src/main/java/persistence/event/PersistEventListener.java
@@ -1,5 +1,0 @@
-package persistence.event;
-
-public interface PersistEventListener extends EventListener {
-    void onPersist(final PersistEvent persistEvent);
-}
