+++ src/test/java/persistence/event/DefaultEventListenerTest.java
@@ -1,6 +1,6 @@
 package persistence.event;
 
-import domain.FixtureEntity.Person;
+import domain.FixtureEntity;
 import mock.MockDmlGenerator;
 import mock.MockJdbcTemplate;
 import org.h2.tools.SimpleResultSet;
@@ -21,9 +21,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-class EventListenerRegistryTest {
+class DefaultEventListenerTest {
+
     private static EntityMetadataProvider entityMetadataProvider;
-    private EventListenerRegistry eventListenerRegistry;
+
+    private DefaultEventListener defaultEventListener;
+
     @BeforeAll
     static void beforeAll() {
         entityMetadataProvider = EntityMetadataProvider.from(new EntityScanner(Application.class));
@@ -39,49 +42,43 @@ class EventListenerRegistryTest {
         rs.addRow(1L, "min", 30, "jongmin4943@gmail.com");
         final EntityPersisters entityPersisters = new EntityPersisters(entityMetadataProvider, new MockDmlGenerator(), new MockJdbcTemplate(rs));
         final EntityLoaders entityLoaders = new EntityLoaders(entityMetadataProvider, new MockDmlGenerator(), new MockJdbcTemplate(rs));
-        eventListenerRegistry = new EventListenerRegistry(new ActionQueue(), entityPersisters, entityLoaders);
+        defaultEventListener = new DefaultEventListener(new ActionQueue(), entityPersisters, entityLoaders);
     }
 
-
     @Test
-    @DisplayName("eventListenerRegistry 의 Persist 로 entity 를 저장할 수 있다.")
-    void eventListenerRegistryPersistTest() {
-        final EventListenerGroup<PersistEventListener> listener = eventListenerRegistry.getListener(EventType.PERSIST);
-        final Person entity = new Person("종민", 30, "jongmin4943@gmail.com");
+    @DisplayName("onPersist의 PersisEvent 정보를 이용해 entity 를 저장할 수 있다.")
+    void onPersistTest() {
+        final FixtureEntity.Person entity = new FixtureEntity.Person("종민", 30, "jongmin4943@gmail.com");
 
-        listener.fireEvent(new PersistEvent(entity), PersistEventListener::onPersist);
+        defaultEventListener.onPersist(new PersistEvent(entity));
 
         assertThat(entity.getId()).isEqualTo(1L);
     }
 
     @Test
-    @DisplayName("eventListenerRegistry 의 MERGE 로 entity 를 변경할 수 있다.")
-    void eventListenerRegistryMergeTest() {
-        final EventListenerGroup<MergeEventListener> listener = eventListenerRegistry.getListener(EventType.MERGE);
-        final Person entity = new Person("종민", 30, "jongmin4943@gmail.com");
+    @DisplayName("onMerge의 MergeEvent 정보를 이용해 entity 를 변경할 수 있다.")
+    void onMergeTest() {
+        final FixtureEntity.Person entity = new FixtureEntity.Person("종민", 30, "jongmin4943@gmail.com");
 
         assertDoesNotThrow(() ->
-            listener.fireEvent(new MergeEvent(entity), MergeEventListener::onMerge)
+                defaultEventListener.onMerge(new MergeEvent(entity))
         );
     }
 
     @Test
-    @DisplayName("eventListenerRegistry 의 DELETE 로 entity 를 삭제할 수 있다.")
-    void eventListenerRegistryDeleteTest() {
-        final EventListenerGroup<DeleteEventListener> listener = eventListenerRegistry.getListener(EventType.DELETE);
-        final Person entity = new Person("종민", 30, "jongmin4943@gmail.com");
+    @DisplayName("onDelete의 DeleteEvent 정보를 이용해 entity 를 변경할 수 있다.")
+    void onDeleteTest() {
+        final FixtureEntity.Person entity = new FixtureEntity.Person("종민", 30, "jongmin4943@gmail.com");
 
         assertDoesNotThrow(() ->
-            listener.fireEvent(new DeleteEvent(entity), DeleteEventListener::onDelete)
+                defaultEventListener.onDelete(new DeleteEvent(entity))
         );
     }
 
     @Test
-    @DisplayName("eventListenerRegistry 의 LOAD 로 entity 를 조회할 수 있다.")
-    void eventListenerRegistryLoadTest() {
-        final EventListenerGroup<LoadEventListener> listener = eventListenerRegistry.getListener(EventType.LOAD);
-
-        final Person entity = listener.fireEventReturn(new LoadEvent<>(1L, Person.class), LoadEventListener::onLoad);
+    @DisplayName("onLoad 의 LoadEvent 정보를 이용해 entity 를 조회할 수 있다.")
+    void onLoadTest() {
+        final FixtureEntity.Person entity = defaultEventListener.onLoad(new LoadEvent<>(1L, FixtureEntity.Person.class));
 
         assertSoftly(softly->{
             softly.assertThat(entity.getId()).isEqualTo(1L);
